% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROO_signatures.R
\name{cleanObject}
\alias{cleanObject}
\title{To be added in the package}
\usage{
cleanObject(merged_obj, colname, expected_labels = "", verbose = FALSE)
}
\arguments{
\item{merged_obj}{a merged object}

\item{colname}{column of interest}
}
\value{
a merged object of smaller size, with only the rows in which the column of interest is part of the binary classification
}
\description{
To be added in the package
}
