% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROO_signatures.R
\name{comp_lda}
\alias{comp_lda}
\title{It computes the LDA for categorical responses (??? need to change? they are actually the predictors)}
\usage{
comp_lda(x, indices_response)
}
\description{
It computes the LDA for categorical responses (??? need to change? they are actually the predictors)
}
