% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROO_signatures.R
\name{comp_logistic}
\alias{comp_logistic}
\title{Performs logistic regression, as explained in the reference}
\usage{
comp_logistic(compData, binaryLabels, relax_binary_assumtion = FALSE)
}
\arguments{
\item{m_obj}{merged_compositional object}

\item{col_idx}{name or index of column of interest in metadata(m_obj)}
}
\description{
Performs logistic regression, as explained in the reference
}
\references{
Van den Boogaart, K. Gerald, and Raimon Tolosana-Delgado. Analyzing compositional data with R. Vol. 122. Berlin: Springer, 2013.
}
