% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{createBarplot}
\alias{createBarplot}
\title{error due to it not being a matrix:  
   No id variables; using all as measure variables
   Rerun with Debug
   Error in `[.data.frame`(.mat, , "Var1") : undefined columns selected \\
(use tomatrix())}
\usage{
createBarplot(
  matrix_exposures,
  angle_rotation_axis = 0,
  order_labels = NULL,
  remove_labels = FALSE,
  levels_signatures = NULL,
  includeMelt = NULL,
  Melt = NULL,
  verbose = TRUE,
  arg_title = "Signature",
  reorder_sigs = NULL
)
}
\description{
error due to it not being a matrix:  
   No id variables; using all as measure variables
   Rerun with Debug
   Error in `[.data.frame`(.mat, , "Var1") : undefined columns selected \\
(use tomatrix())
}
