% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROO_signatures.R
\name{createDendrogram}
\alias{createDendrogram}
\title{Create a dendrogram, using Aithchison distance, of the samples in a merged
object. The labels are coloured according to the one of the columns of its
metadata dataframe (to be specified in name_clinical). There is also an
option (bool_comparison) to add a second dendrogram using Euclidean distance,
for comparison.
WARNING! colours for the second dendrogram need fixing}
\usage{
createDendrogram(merged_object, name_clinical, bool_comparison, ...)
}
\description{
Create a dendrogram, using Aithchison distance, of the samples in a merged
object. The labels are coloured according to the one of the columns of its
metadata dataframe (to be specified in name_clinical). There is also an
option (bool_comparison) to add a second dendrogram using Euclidean distance,
for comparison.
WARNING! colours for the second dendrogram need fixing
}
