% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROO_signatures.R
\name{deprecated_comp_logistic}
\alias{deprecated_comp_logistic}
\title{Perform logistic regression}
\usage{
deprecated_comp_logistic(merged_obj, colname)
}
\arguments{
\item{merged_obj}{a merged_compositional object}

\item{colname}{either the name or the index of the column in metadata(merged_compositional) with which to perform logistic regression}
}
\value{
The summary of a logistic regression
}
\description{
Perform logistic regression
}
