% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{give_x_matrix}
\alias{give_x_matrix}
\title{Give a X (covariate) matrix for a scenario in which each patient contributes to two samples, and the samples are sorted by group, then by patient.}
\usage{
give_x_matrix(n_times_2)
}
\arguments{
\item{n_times_2:}{the number of samples times two}
}
\value{
the X matrix
}
\description{
Give a X (covariate) matrix for a scenario in which each patient contributes to two samples, and the samples are sorted by group, then by patient.
}
