% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{give_z_matrix}
\alias{give_z_matrix}
\title{Give a Z matrix for a scenario in which each patient contributes to two samples, and the samples are sorted by group, then by patient.}
\usage{
give_z_matrix(n_times_2)
}
\arguments{
\item{n_times_2:}{the number of samples times two}
}
\value{
the Z matrix
}
\description{
Give a Z matrix for a scenario in which each patient contributes to two samples, and the samples are sorted by group, then by patient.
}
