% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA_functions.R
\name{L_to_cov_wrapper_TMB}
\alias{L_to_cov_wrapper_TMB}
\title{Give the estimated covariance matrix from the L matrix (see https://kaskr.github.io/adcomp/classdensity_1_1UNSTRUCTURED__CORR__t.html)}
\usage{
L_to_cov_wrapper_TMB(TMB_obj, name_cov_par = "cov_RE", d = NULL)
}
\arguments{
\item{TMB_obj:}{TMB object}

\item{name_cov_par:}{the name of the parameters which correspond to the elements of L}
}
\value{
the covariance matrix
}
\description{
Give the estimated covariance matrix from the L matrix (see https://kaskr.github.io/adcomp/classdensity_1_1UNSTRUCTURED__CORR__t.html)
}
