% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROO_functions.R
\name{load_PCAWG}
\alias{load_PCAWG}
\title{Function to load an exposures object}
\usage{
load_PCAWG(
  ct,
  typedata,
  simulation = FALSE,
  path_to_data = "../../data/",
  read_directly = FALSE,
  old_version_creating_X_Z = F,
  load_all_sigs = F,
  override_warning_X_Z = F
)
}
\arguments{
\item{ct:}{cancer type, if the data are organised as in the PCAWG dataset. Otherwise, name of ile}

\item{typedata:}{type of data (simulation/nucleotidesubstitution1/nucleotidesubstitution3/signatures/etc.), if the data are organised as in the PCAWG dataset}

\item{simulation:}{boolean indicating if the data are simulations (i.e. not organised as in the PCAWG dataset)}

\item{path_to_data:}{path to folder which includes the file to upload}

\item{read_directly:}{boolean indicating if <ct> includes the filename}

\item{old_version_creating_X_Z:}{boolean indicating if the input file contains exposures organised as in <<Display A>>}

\item{load_all_sigs:}{load all signatures (otherwise, only active signatures are loaded)}

\item{override_warning_X_Z:}{allow X and Z to be generated assuming that the input file contains exposures organised as in <<Display A>>}
}
\value{
the covariance matrix
get the first element of the list, because the second is the
set of the parameters used in the creation of the dataset
}
\description{
Function to load an exposures object
}
