% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROO_signatures.R
\docType{class}
\name{merged_compositional-class}
\alias{merged_compositional-class}
\title{Class whiuch inludes features of both metadata and
 mutational signatures.
 types_metadata is a vector indicating which type of data each
 column in df (metadata) is.}
\description{
Class whiuch inludes features of both metadata and
 mutational signatures.
 types_metadata is a vector indicating which type of data each
 column in df (metadata) is.
}
