% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROO_signatures.R
\name{merged_compositional_to_sign}
\alias{merged_compositional_to_sign}
\title{Convert a 'sign' object to a 'merged_compositional_to_sign'
(note loss of information)}
\usage{
merged_compositional_to_sign(x)
}
\description{
Convert a 'sign' object to a 'merged_compositional_to_sign'
(note loss of information)
}
