% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{softmax}
\alias{softmax}
\title{Compute the softmax-transformed version of a vector or a matrix. Reminder: if x is an ALR-transformed compositional vector, you want to append 0 as the final element of the vector (if x is a vector), or a final column of zeros (if x is a matrix)}
\usage{
softmax(x)
}
\arguments{
\item{x:}{a vector or a matrix}
}
\value{
the softmax-transformed x
}
\description{
Compute the softmax-transformed version of a vector or a matrix. Reminder: if x is an ALR-transformed compositional vector, you want to append 0 as the final element of the vector (if x is a vector), or a final column of zeros (if x is a matrix)
}
