% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{vector_to_ct_list}
\alias{vector_to_ct_list}
\title{given a vector which contains two or more types of features for the same cancer types,
convert it to a matrix with the pairing per cancer type}
\usage{
vector_to_ct_list(vec)
}
\description{
given a vector which contains two or more types of features for the same cancer types,
convert it to a matrix with the pairing per cancer type
}
