% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA_functions.R
\name{wald_TMB_wrapper}
\alias{wald_TMB_wrapper}
\title{Generalised wald test for the slope beta values}
\usage{
wald_TMB_wrapper(i, verbatim = TRUE, fail_non_converged = T)
}
\arguments{
\item{i:}{TMB output object}

\item{verbatim:}{verbatim boolean}

\item{fail_non_converged:}{should a non-converged run lead to an NA value?}
}
\description{
Generalised wald test for the slope beta values
}
