% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA_functions.R
\name{wrapper_run_TMB}
\alias{wrapper_run_TMB}
\title{Function to run the model using TMB}
\usage{
wrapper_run_TMB(
  model,
  object = NULL,
  smart_init_vals = T,
  use_nlminb = F,
  initial_params = NULL
)
}
\arguments{
\item{model:}{name of model: Options: {fullRE_M, diagRE_M, FE_DM, fullRE_DM, fullRE_DMonefixedlambda, diagRE_DM, fullREDMsinglelambda, fullREDMsinglelambda, singleRE_DM, fullREhalfDM, diagREDMsinglelambda, FEDMsinglelambda, fullRE_multinomial_REv2, diagDMpatientlambda, fullDMpatientlambda}}

\item{object:}{object of type exposures_inputTMB (including X, Y, Z). See <simplified_object> as an example.}

\item{smart_init_vals:}{boolean, whether a fixed-effects multinomial regression should be run first to get initial estimates}

\item{use_nlminb:}{boolean, whether <nlminb> should be used for estimation. Alternatively, <optim> is used}

\item{initial_params:}{(optional) list of initial parameters for estimation}
}
\description{
Function to run the model using TMB
}
