% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROO_signatures.R
\name{count_matrix}
\alias{count_matrix}
\title{Retrieve count matrix from object of class sign.}
\usage{
count_matrix(X)
}
\arguments{
\item{X}{object of class sign}
}
\description{
Retrieve count matrix from object of class sign.
}
